refactor(server/events): extract makeAnnotationsObserver to observers/annotations.ts
Move the annotations Y.Map observer body into a standalone factory function.
attachObservers pushes the returned cleanup. Local annotationsMap declaration
retained in attachObservers for the still-inline replies observer.
